# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CareerPilot \u2014 Multi-agent AI job discovery pipeline",
    "description": "Three agents (Scout finds company job boards, Analyst scores postings against my resume, Coach identifies recurring skill gaps), LangGraph orchestration, RAG over the job archive, and a hand-labeled evaluation set.\n\nThis is the HTTP layer over that pipeline. It wraps adapters/, filters.py, agents/, orchestrator.py, db.py, extraction.py and rag.py - it does not reimplement any of them.\n\n**The system finds and ranks; the human applies.** Nothing here submits an application to an employer, and no endpoint exists that could - that is a design decision, not a missing feature.\n\nRanking quality is measured against the hand-labeled set rather than assumed. The current result is marginal - stage-1 MRR sits only slightly above the random baseline on a small overlap. See GET /api/meta/evaluation for the numbers and their caveats.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/companies": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "List Companies",
        "operationId": "list_companies_api_companies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompaniesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "companies"
        ],
        "summary": "Add Company",
        "description": "Appends to companies.yaml.\n\nWritten with yaml.safe_dump over the parsed list rather than appending\na hand-built string, so a company name containing a colon or a quote\ncan't produce a file that no longer parses - the same reasoning\nagents/scout.py's _found_as_yaml already applies.",
        "operationId": "add_company_api_companies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCompanyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddCompanyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/companies/scout": {
      "post": {
        "tags": [
          "companies"
        ],
        "summary": "Run Scout",
        "description": "Runs the Scout agent against one company name.\n\nBlocking and potentially slow - Scout probes up to MAX_TOTAL_ATTEMPTS\nboard endpoints and may consult the LLM once when mechanical candidates\nare exhausted. Kept synchronous rather than pushed through the run\nmanager because it's bounded, single-company, and the caller wants the\nanswer, not a stream. Writes nothing.",
        "operationId": "run_scout_api_companies_scout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/coach/skill-gaps": {
      "get": {
        "tags": [
          "coach"
        ],
        "summary": "Skill Gaps",
        "operationId": "skill_gaps_api_coach_skill_gaps_get",
        "parameters": [
          {
            "name": "threshold",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 0,
              "default": 40,
              "title": "Threshold"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillGapsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/coach/ask": {
      "post": {
        "tags": [
          "coach"
        ],
        "summary": "Ask",
        "description": "A market-gap question, answered over retrieved JDs.\n\nSpends one LLM call. Imports agents.coach's market_gap and the LLM\nclient lazily so that merely importing the API doesn't pull the RAG\nstack (which reaches ranking.py, and through it sentence-transformers)\ninto every process that serves a job list.",
        "operationId": "ask_api_coach_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddCompanyRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ats": {
            "type": "string",
            "title": "Ats"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "ats",
          "token"
        ],
        "title": "AddCompanyRequest"
      },
      "AddCompanyResponse": {
        "properties": {
          "added": {
            "type": "boolean",
            "title": "Added"
          },
          "company": {
            "$ref": "#/components/schemas/CompanyStats"
          },
          "total_companies": {
            "type": "integer",
            "title": "Total Companies"
          }
        },
        "type": "object",
        "required": [
          "added",
          "company",
          "total_companies"
        ],
        "title": "AddCompanyResponse"
      },
...
      "AskRequest": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "k": {
            "type": "integer",
            "title": "K",
            "default": 12
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AskRequest"
      },
      "AskResponse": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "k": {
            "type": "integer",
            "title": "K"
          },
          "retrieved": {
            "items": {
              "$ref": "#/components/schemas/RetrievedChunk"
            },
            "type": "array",
            "title": "Retrieved"
          },
          "pool_size": {
            "type": "integer",
            "title": "Pool Size"
          },
          "retrieved_count": {
            "type": "integer",
            "title": "Retrieved Count"
          },
          "retrieval_was_noop": {
            "type": "boolean",
            "title": "Retrieval Was Noop"
          },
          "population_size": {
            "type": "integer",
            "title": "Population Size"
          }
        },
        "type": "object",
        "required": [
          "question",
          "answer",
          "k",
          "pool_size",
          "retrieved_count",
          "retrieval_was_noop",
          "population_size"
        ],
        "title": "AskResponse",
        "description": "A Coach answer plus what it actually retrieved. The retrieved set is\nreturned alongside the answer, not hidden, so the answer can be checked\nagainst its own evidence rather than taken on trust - the whole point of\nRAG over a chatbot.\n\n`retrieval_was_noop` is surfaced, not buried: at this project's current\ncorpus size, k frequently exceeds the matching pool, meaning \"retrieval\"\nreturned everything and ranked nothing. An answer built that way is\nstill useful, but calling it retrieval-augmented would overstate what\nhappened. See rag.RetrievalResult.is_noop."
      },
...
      "CompaniesResponse": {
        "properties": {
          "companies": {
            "items": {
              "$ref": "#/components/schemas/CompanyStats"
            },
            "type": "array",
            "title": "Companies"
          },
          "zero_survivor_companies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Zero Survivor Companies"
          }
        },
        "type": "object",
        "required": [
          "companies",
          "zero_survivor_companies"
        ],
        "title": "CompaniesResponse"
      },
      "CompanyStats": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ats": {
            "type": "string",
            "title": "Ats"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "jobs_fetched": {
            "type": "integer",
            "title": "Jobs Fetched"
          },
          "survivors": {
            "type": "integer",
            "title": "Survivors"
          }
        },
        "type": "object",
        "required": [
          "name",
          "ats",
          "token",
          "jobs_fetched",
          "survivors"
        ],
        "title": "CompanyStats",
        "description": "One companies.yaml entry plus what it actually contributed.\nfetched > 0 with survivors == 0 means the company is costing a nightly\nfetch and returning nothing usable - a removal candidate."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RetrievedChunk": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "text",
          "score"
        ],
        "title": "RetrievedChunk",
        "description": "One retrieved JD requirements excerpt and its cosine similarity to\nthe question."
      },
...
      "ScoutEmptyBoard": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "token",
          "source"
        ],
        "title": "ScoutEmptyBoard"
      },
      "ScoutRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ScoutRequest"
      },
      "ScoutResponse": {
        "properties": {
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "ats": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ats"
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "job_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Count"
          },
          "conclusion": {
            "type": "string",
            "title": "Conclusion"
          },
          "requests_used": {
            "type": "integer",
            "title": "Requests Used"
          },
          "empty_boards": {
            "items": {
              "$ref": "#/components/schemas/ScoutEmptyBoard"
            },
            "type": "array",
            "title": "Empty Boards"
          }
        },
        "type": "object",
        "required": [
          "company_name",
          "success",
          "conclusion",
          "requests_used"
        ],
        "title": "ScoutResponse",
        "description": "What Scout concluded. `success=False` with a non-empty\n`empty_boards` is a distinct, useful outcome from a flat miss: a real\nboard exists, it just has zero open jobs right now, so it's worth\nre-scouting later rather than writing the company off.\n\nNothing here is written to companies.yaml - POST /api/companies does\nthat, deliberately as a separate, explicit step."
      },
      "SkillGap": {
        "properties": {
          "skill": {
            "type": "string",
            "title": "Skill"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "skill",
          "count"
        ],
        "title": "SkillGap"
      },
      "SkillGapsResponse": {
        "properties": {
          "threshold": {
            "type": "integer",
            "title": "Threshold"
          },
          "job_count": {
            "type": "integer",
            "title": "Job Count"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillGap"
            },
            "type": "array",
            "title": "Skills"
          }
        },
        "type": "object",
        "required": [
          "threshold",
          "job_count"
        ],
        "title": "SkillGapsResponse",
        "description": "Missing-skills frequency across stage-1-scored jobs below a\nthreshold. Stage-1 only, deliberately: mixing stage-1 and stage-2 rows\nwould aggregate judgments made by two different models into one count\nthat means neither. See agents/coach.py."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}